Update geometry stuff

+++ ros2/src/vtr_path_tracker/include/vtr_path_tracker/robust_mpc/mpc/utilities.h
@@ -9,13 +9,14 @@ namespace utils {
 /**
  * @brief thetaWrap Wrap an angle to (-pi, pi]
  * @param th_in The angle
+ * @note  Assumes abs(th_in) < 3pi
  * @return The equivalent angle between (-pi, pi]
  */
-float thetaWrap(float th_in);
+double thetaWrap(double th_in);
 
 /** @brief Returns the sign of a number as a float (-1.,1.)
  */
-float getSign(float number);
+double getSign(double number);
 
 } // utils
 } // path_tracker
