+++ ros2/src/vtr_path_tracker/src/robust_mpc/mpc/utilities.cpp
@@ -4,8 +4,8 @@ namespace vtr {
 namespace path_tracker {
 namespace utils {
 
-float thetaWrap(float th_in) {
-  float th_out = th_in;
+double thetaWrap(double th_in) {
+  double th_out = th_in;
   if (th_in > M_PI) {
     th_out = th_in - 2 * M_PI;
   } else if (th_in < -M_PI) {
@@ -14,8 +14,8 @@ float thetaWrap(float th_in) {
   return th_out;
 }
 
-float getSign(float number) {
-  return (float) ((0.0 < number) - (number < 0.0));
+double getSign(double number) {
+  return (double) ((0.0 < number) - (number < 0.0));
 }
 
 } // utils
